move code to types
