+++ src/ext/openai/index.ts
@@ -1,2 +1,3 @@
 export * from './openai';
 export * from './prompt';
+export * from './types';
